move device functions to device.c
