+++ ext/mri/arrayfire.c
@@ -544,115 +544,7 @@ static VALUE arf_set_native_id(VALUE self){
 
 // Device
 
-static VALUE arf_info(VALUE self){
-  af_info();
-  return Qnil;
-}
-
-static VALUE arf_init2(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_info_string(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_device_info(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_get_device_count(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_get_dbl_support(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_set_device(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_get_device(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_sync(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_alloc_device(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_free_device(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_alloc_pinned(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_free_pinned(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_alloc_host(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_free_host(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_device_array(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_device_mem_info(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_print_mem_info(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_device_gc(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_set_mem_step_size(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_get_mem_step_size(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_lock_device_ptr(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_unlock_device_ptr(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_lock_array(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_unlock_array(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_is_locked_array(VALUE self){
-  return Qnil;
-}
-
-static VALUE arf_get_device_ptr(VALUE self){
-  return Qnil;
-}
-
+#include "cmodules/device.c"
 #include "cmodules/opencl.c"
 #include "cmodules/data.c"
 #include "cmodules/lapack.c"
